CI-338 moved all repositories into repositories folder
