+++ src/Web/Sfa.Das.Sas.Infrastructure/DependencyResolution/InfrastructureRegistry.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Diagnostics;
 using System.Reflection;
+using Sfa.Das.Sas.Infrastructure.Repositories;
 
 namespace Sfa.Das.Sas.Infrastructure.DependencyResolution
 {
