+++ src/Web/Sfa.Das.Sas.Infrastructure/Repositories/ApprenticeshipProviderApiRepository.cs
@@ -6,9 +6,10 @@ using Sfa.Das.Sas.ApplicationServices.Models;
 using Sfa.Das.Sas.Core.Configuration;
 using Sfa.Das.Sas.Core.Domain.Model;
 using Sfa.Das.Sas.Core.Domain.Services;
+using Sfa.Das.Sas.Infrastructure.Elasticsearch;
 using SFA.DAS.NLog.Logger;
 
-namespace Sfa.Das.Sas.Infrastructure.Elasticsearch
+namespace Sfa.Das.Sas.Infrastructure.Repositories
 {
     public sealed class ApprenticeshipProviderApiRepository : IApprenticeshipProviderRepository
     {
