+++ src/Web/Sfa.Das.Sas.Infrastructure/Repositories/StandardApiRepository.cs
@@ -1,18 +1,19 @@
 ﻿
-namespace Sfa.Das.Sas.Infrastructure.Elasticsearch
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using Nest;
-    using SFA.DAS.Apprenticeships.Api.Client;
-    using SFA.DAS.Apprenticeships.Api.Types.Exceptions;
-    using Sfa.Das.Sas.ApplicationServices.Models;
-    using Sfa.Das.Sas.Core.Configuration;
-    using Sfa.Das.Sas.Core.Domain.Model;
-    using Sfa.Das.Sas.Core.Domain.Services;
-    using Sfa.Das.Sas.Infrastructure.Mapping;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Nest;
+using Sfa.Das.Sas.ApplicationServices.Models;
+using Sfa.Das.Sas.Core.Configuration;
+using Sfa.Das.Sas.Core.Domain.Model;
+using Sfa.Das.Sas.Core.Domain.Services;
+using Sfa.Das.Sas.Infrastructure.Elasticsearch;
+using Sfa.Das.Sas.Infrastructure.Mapping;
+using SFA.DAS.Apprenticeships.Api.Client;
+using SFA.DAS.Apprenticeships.Api.Types.Exceptions;
 
+namespace Sfa.Das.Sas.Infrastructure.Repositories
+{
     public sealed class StandardApiRepository : IGetStandards
     {
         private readonly IElasticsearchCustomClient _elasticsearchCustomClient;
